+++ public/html/ren_crawler.pug
@@ -3,9 +3,9 @@ html
   head
     title Cornerth.
     meta(charSet="utf-8")
-    meta(name="robots", content="nofollow")
-    meta(name='title', content= '#{title}')
-    meta(name='description', content= '#{descrip}')
-    meta(property='og:url', content= '#{ogurl}')
-    meta(property='og:image', content= '#{ogimg}')
+    meta(name="robots" content="nofollow")
+    meta(name='title' content= title)
+    meta(name='description' content= descrip)
+    meta(property='og:url' content= ogurl)
+    meta(property='og:image' content= ogimg)
 body
